use size_t in shuffle_array

+++ src/redo.c
@@ -209,8 +209,8 @@ cleanup_dbfile:
 void
 shuffle_array(unsigned int seed, char *arr[], size_t len) {
 	srand(seed);
-	for(int i = len - 1; i >= 1; i--) {
-		int j = (rand() % i);
+	for(size_t i = len - 1; i >= 1; i--) {
+		size_t j = (rand() % i);
 		char *tmp = arr[i];
 		arr[i] = arr[j];
 		arr[j] = tmp;
